refactor(swap): move api-maker file
